# api/swagger/gateway/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Платежный шлюз для E-commerce системы",
        "contact": {
            "name": "Курмыза Павел",
            "email": "tmrrwnxtsn@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/payout/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Выплаты"
                ],
                "summary": "Создать запрос на вывод средств",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.payoutCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.payoutCreateResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/payout/methods": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Выплаты"
                ],
                "summary": "Получить список способов для вывода средств",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор клиента",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Идентификатор сессии клиента",
                        "name": "session_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Валюта выплаты в соответствии со стандартом ISO 4217",
                        "name": "currency",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Код языка, обозначение по RFC 5646",
                        "name": "lang_code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.payoutMethodsResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
        "/payout/{id}/confirm": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Выплаты"
                ],
                "summary": "Подтвердить запрос на вывод средств",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Идентификатор операции",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.payoutConfirmRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "schema": {
                            "$ref": "#/definitions/v1.payoutConfirmResponse"
                        }
                    },
                    "default": {
                        "description": "Ответ с ошибкой",
                        "schema": {
                            "$ref": "#/definitions/v1.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "v1.commission": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "absolute": {
                    "description": "Значение комиссии (передается, если тип комиссии, \"type\", равен \"fixed\" или \"combined\")",
                    "type": "number",
                    "example": 10
                },
                "caption": {
                    "description": "Текстовая репрезентация комиссии (передается, если тип комиссии, \"type\", равен \"text\")",
                    "type": "string",
                    "example": "Комиссия взимается провайдером"
                },
                "currency": {
                    "description": "Код валюты комиссии (передается, если тип комиссии, \"type\", равен \"fixed\" или \"combined\")",
                    "type": "string",
                    "example": "RUB"
                },
                "percent": {
                    "description": "Значение комиссии (передается, если тип комиссии, \"type\", равен \"percent\" или \"combined\")",
                    "type": "number",
                    "example": 11.99
                },
                "type": {
                    "description": "Тип комиссии:\n* \"percent\" - комиссия в процентах\n* \"fixed\" - фиксированная комиссия\n* \"combined\" - комбинированная комиссия, из процентов и фиксированной суммы\n* \"text\" - текстовая коммиссия, например, \"Взимается провайдером\"",
                    "type": "string",
                    "example": "combined"
                }
            }
        },
        "v1.errorContent": {
            "type": "object",
            "required": [
                "code",
                "description"
            ],
            "properties": {
                "code": {
                    "description": "Код ошибки",
                    "type": "string",
                    "example": "InvalidRequest"
                },
                "description": {
                    "description": "Описание ошибки для разработки",
                    "type": "string",
                    "example": "user_id param is required"
                },
                "message": {
                    "description": "Сообщение об ошибке для клиента",
                    "type": "string",
                    "example": "Internal server error occurred. Please try again later."
                }
            }
        },
        "v1.errorResponse": {
            "type": "object",
            "required": [
                "error",
                "success"
            ],
            "properties": {
                "error": {
                    "description": "Развернутая информация об ошибке",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.errorContent"
                        }
                    ]
                },
                "success": {
                    "description": "Результат обработки запроса (всегда false)",
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "v1.limits": {
            "type": "object",
            "required": [
                "currency",
                "max_amount",
                "min_amount"
            ],
            "properties": {
                "currency": {
                    "description": "Код валюты лимита в соответствии со стандартом ISO 4217",
                    "type": "string",
                    "example": "RUB"
                },
                "max_amount": {
                    "description": "Максимальное значение суммы (в дробных единицах)",
                    "type": "number",
                    "example": 60000
                },
                "min_amount": {
                    "description": "Минимальное значение суммы (в дробных единицах)",
                    "type": "number",
                    "example": 100
                }
            }
        },
        "v1.method": {
            "type": "object",
            "required": [
                "commission",
                "external_method",
                "external_system",
                "favorite",
                "id",
                "limits",
                "name"
            ],
            "properties": {
                "commission": {
                    "description": "Объект, содержащий данные о комиссии",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.commission"
                        }
                    ]
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "external_system": {
                    "description": "Внутренний код платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa"
                },
                "favorite": {
                    "description": "Флаг о том, что платежная система добавлена в избранное",
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "description": "Идентификатор ПС из внутреннего справочника",
                    "type": "string",
                    "example": "CARD"
                },
                "limits": {
                    "description": "Массив объектов, содержащих данные о лимитах",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.limits"
                    }
                },
                "name": {
                    "description": "Название платежной системы",
                    "type": "string",
                    "example": "Банковская карта"
                },
                "tools": {
                    "description": "Массив объектов, содержащих данные о сохраненных платежных средствах",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.tool"
                    }
                }
            }
        },
...
        "v1.payoutConfirmRequest": {
            "type": "object",
            "required": [
                "confirmation_code",
                "lang_code",
                "session_id",
                "user_id"
            ],
            "properties": {
                "confirmation_code": {
                    "description": "Код подтверждения вывода средств",
                    "type": "string",
                    "example": "123456"
                },
                "lang_code": {
                    "description": "Код языка, обозначение по RFC 5646",
                    "type": "string",
                    "example": "en"
                },
                "session_id": {
                    "description": "Идентификатор сессии клиента",
                    "type": "string",
                    "example": "LRXZmXPGusPCfys48LadjFew"
                },
                "user_id": {
                    "description": "Идентификатор клиента",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.payoutConfirmResponse": {
            "type": "object",
            "required": [
                "message",
                "success"
            ],
            "properties": {
                "message": {
                    "description": "Сообщение, которое необходимо показать клиенту",
                    "type": "string",
                    "example": "Вывод средств подтвержден."
                },
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "v1.payoutCreateRequest": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "external_method",
                "external_system",
                "lang_code",
                "session_id",
                "tool_id",
                "user_id"
            ],
            "properties": {
                "additional_data": {
                    "description": "Дополнительная информация, специфичная для платежной системы, к которой направляется целевой запрос",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "ip": "127.0.0.1",
                        "phone_number": "+71234567890"
                    }
                },
                "amount": {
                    "description": "Сумма выплаты в минорных единицах валюты (копейки, центы и т.п.)",
                    "type": "integer",
                    "minimum": 100,
                    "example": 10000
                },
                "currency": {
                    "description": "Валюта выплаты в соответствии со стандартом ISO 4217",
                    "type": "string",
                    "example": "RUB"
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "external_system": {
                    "description": "Внутренний код платежной системы, к которой направляется целевой запрос",
                    "type": "string",
                    "example": "yookassa"
                },
                "lang_code": {
                    "description": "Код языка, обозначение по RFC 5646",
                    "type": "string",
                    "example": "en"
                },
                "session_id": {
                    "description": "Идентификатор сессии клиента",
                    "type": "string",
                    "example": "LRXZmXPGusPCfys48LadjFew"
                },
                "tool_id": {
                    "description": "Идентификатор сохраненного платежного средства",
                    "type": "string",
                    "example": "2dc32aa0-000f-5000-8000-16d7bc6cd09f"
                },
                "user_id": {
                    "description": "Идентификатор клиента",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "v1.payoutCreateResponse": {
            "type": "object",
            "required": [
                "operation_id",
                "success"
            ],
            "properties": {
                "operation_id": {
                    "description": "Идентификатор созданной выплаты",
                    "type": "integer",
                    "example": 1
                },
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "v1.payoutMethodsResponse": {
            "type": "object",
            "required": [
                "payout_methods",
                "success"
            ],
            "properties": {
                "payout_methods": {
                    "description": "Массив платежных методов, доступных для вывода средств",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/v1.method"
                    }
                },
                "success": {
                    "description": "Результат обработки запроса (всегда true)",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "v1.tool": {
            "type": "object",
            "required": [
                "caption",
                "external_method",
                "id",
                "name",
                "type"
            ],
            "properties": {
                "caption": {
                    "description": "Значение платежного средства, например:\n* Маскированная банковская карта\n* Номер электронного кошелька\n* Адрес электронной почты\n* и т.д.",
                    "type": "string",
                    "example": "444444******4444"
                },
                "details": {
                    "description": "Дополнительная информация о платежном средстве",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.toolDetails"
                        }
                    ]
                },
                "external_method": {
                    "description": "Внутренний код платежного метода платежной системы, к которой относится платежное средство",
                    "type": "string",
                    "example": "yookassa_bank_card"
                },
                "id": {
                    "description": "Идентификатор платежного средства",
                    "type": "string",
                    "example": "2dc32aa0-000f-5000-8000-16d7bc6cd09f"
                },
                "name": {
                    "description": "Название платежного средства",
                    "type": "string",
                    "example": "Карта брата"
                },
                "type": {
                    "description": "Тип платежного средства:\n* Банковская карта - \"card\"\n* Электронный кошелек - \"wallet\"",
                    "type": "string",
                    "example": "card"
                }
            }
        },
        "v1.toolDetails": {
            "type": "object",
            "properties": {
                "bank_name": {
                    "description": "Название банка, выпустившего банковскую карту",
                    "type": "string",
                    "example": "Sberbank"
                },
                "card_holder": {
                    "description": "Владелец банковской карты",
                    "type": "string",
                    "example": "Ivanov Ivan"
                },
                "card_type": {
                    "description": "Тип банковской карты",
                    "type": "string",
                    "example": "Visa"
                },
                "expiry_month": {
                    "description": "Срок действия банковской карты (месяц, MM)",
                    "type": "integer",
                    "example": 10
                },
                "expiry_year": {
                    "description": "Срок действия банковской карты (год, YYYY)",
                    "type": "integer",
                    "example": 2023
                },
                "wallet_number": {
                    "description": "Номер электронного кошелька",
                    "type": "string",
                    "example": "410011758831136"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Секретный ключ",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}